move to_color_strings to _engine_utils
